Changed progress listener worked() to use double

+++ bundles/org.eclipse.core.runtime/src/org/eclipse/core/runtime/jobs/IProgressListener.java
@@ -44,5 +44,5 @@ public interface IProgressListener {
 	 * 
 	 * @see IProgressMonitor#worked
 	 */
-	public void worked(Job job, int work);
+	public void worked(Job job, double work);
 }
